requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
pymongo==4.16.0
python-dotenv==1.2.2
yt-dlp>=2025.01.15
aiofiles==25.1.0
